Add report service
