+++ Backend/apps/financials/views.py
@@ -2,7 +2,8 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.renderers import JSONRenderer
 from rest_framework import status
-from .modules import *
+from .services import financial_report_service as report_service
+from .services import financial_statement_service as statement_service
 
 
 class CompanyLookupView(APIView):
@@ -12,14 +13,15 @@ class CompanyLookupView(APIView):
         query = request.query_params["query"]
         consolidation = request.query_params["consolidation"]
 
-        report = FinancialReport()
-        corporate_code = report.get_corporate_code(query)
+        financial_report_service = report_service.FinancialReportService()
+        financial_statement_service = statement_service.FinancialStatementService()
+        corporate_code = financial_report_service.get_corporate_code(query)
 
-        if str(corporate_code) == CorporateCodeNotFoundError:
-            return Response(data={"message": CorporateCodeNotFoundError}, status=status.HTTP_404_NOT_FOUND)
-        financial_statements = report.get_financial_statements(
-            corporate_code, consolidation_key)
-        financial_reports = report.get_financial_reports(
+        if corporate_code == report_service.CorporateCodeNotFoundError:
+            return Response(data={"message": report_service.CorporateCodeNotFoundError}, status=status.HTTP_404_NOT_FOUND)
+        financial_statements = financial_statement_service.get_financial_statements(
+            corporate_code, consolidation)
+        financial_reports = financial_report_service.get_financial_reports(
             financial_statements)
 
         return Response(data={"reports: ": financial_reports}, status=status.HTTP_200_OK)
